+++ docs/docs.r
@@ -2,7 +2,7 @@
 # Configuration & R packages for documents --------------------------------
 # -------------------------------------------------------------------------
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date   .: 2021-04-28
+#> date   .: 2021-07-17
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -29,7 +29,7 @@ remove(cran, pkg)
 knitr::opts_chunk$set(
     fig.align = "center" # Center images in the export file
   , out.width = "98%" # Figure width in html
-  , echo = FALSE # Avoid print code in the export file
+  # , echo = FALSE # Avoid print code in the export file
   , message = FALSE # Avoid print messages in the export file
   , warning = FALSE # Avoid print messages in the export file
   , collapse = TRUE # Collapse text output into source blocks